refactor: `Infrascructure Repository Herance`
